Sync 23-03-2020

+++ project/manager/js/server/set_project.js
@@ -1,19 +1,28 @@
 const fs = require('fs-extra');
 module.exports = {
   responce: (pDir, project) => {
-    /* 
-      Update /app/projects_archive/nuxt.project.js 
-    */
-    let content = `${fs.readFileSync(`${pDir}nuxt.project.js`)}`;
-    const old_project = `${content}`.split(/'/)[1];
-    content = content.replace(old_project, project);
-    fs.writeFileSync(`${pDir}nuxt.project.js`, content);
-    /*
-       Read /app/projects_archive/[new_project]/nuxt.layout.js
-    */
-    let layout = `${fs.readFileSync(`${pDir}${project}/nuxt.layout.js`)}`;
-    layout = `${layout}`.split(/'/)[1];
+    if (!fs.existsSync(`${pDir}${project}`)) {
+      return { project_not_exist: true };
+    } else {
+      /* 
+          Update /app/projects_archive/nuxt.project.js 
+        */
+      content = `export const SOJUZ_PROJECT = '${project}';\n`;
+      fs.writeFileSync(`${pDir}nuxt.project.js`, content);
 
-    return { project, layout };
+      /*
+        Read /app/projects_archive/[new_project]/nuxt.layout.js
+      */
+      let layout = `${fs.readFileSync(`${pDir}${project}/nuxt.layout.js`)}`;
+      layout = `${layout}`.split(/'/)[1];
+
+      /*
+        Read /app/projects_archive/[new_project]/nuxt.css.collection.json
+      */
+      let css_collection = `${fs.readFileSync(`${pDir}${project}/nuxt.css.collection.json`)}`;
+      css_collection = JSON.parse(css_collection);
+
+      return { project, layout, css_collection };
+    }
   },
 };
